Major List Overhaul
Replaced old table based lists with a snippet+flex box system.
Projects & lists now have a 'preview snippet' that gets embedded, instead of a hard coded link. Added a system to hide WIP projects unless there is #wip in the URI. This can be toggled with the link in the bottom left

Added some additional information to ABW's page

Mild refactoring, and a few extra assets